Småfikser før feedback
